# improve performance of BinaryData construction