Avoids using `stream.ToArray()` in favor of ReadOnlyMemory.

| Method  | Mean     | Error     | StdDev    | Ratio | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------- |---------:|----------:|----------:|------:|-------:|-------:|----------:|------------:|
| ToArray | 4.029 us | 0.0377 us | 0.0334 us |  1.00 | 2.9984 | 0.1831 |  24.53 KB |        1.00 |
| ROM     | 3.818 us | 0.0177 us | 0.0148 us |  0.95 | 2.4033 | 0.0839 |  19.66 KB |        0.80 |